lets data mover use cli
